Modify to work with snap-0.10.*
snap-0.10 uses the lens package instead of data-lens, so we use
lens package lenses in M.hs. Also remove some unnecessary imports
in S.hs, bump the snap dependency, and add a direct dependency on
the lens package. (The last is needed because we're using functions
from Control.Lens that Snap does not re-export.)